Data.Deriving.Internal

Description

Template Haskell-related utilities.

Note: this is an internal module, and as such, the API presented here is not guaranteed to be stable, even between minor releases of this library.

Documentation

applySubstitutionKind :: Map Name Kind -> Type -> Type

substNameWithKind :: Name -> Kind -> Type -> Type

substNamesWithKindStar :: [Name] -> Type -> Type

data Via a b infix 0

A type-level modifier intended to be used in conjunction with deriveVia. Refer to the documentation for deriveVia for more details.

fmapConst :: f b -> (a -> b) -> f a -> f b

replaceConst :: f a -> a -> f b -> f a

foldrConst :: b -> (a -> b -> b) -> b -> t a -> b

foldMapConst :: m -> (a -> m) -> t a -> m

nullConst :: Bool -> t a -> Bool

traverseConst :: f (t b) -> (a -> f b) -> t a -> f (t b)

eqConst :: Bool -> a -> a -> Bool

eq1Const :: Bool -> f a -> f a -> Bool

liftEqConst :: Bool -> (a -> b -> Bool) -> f a -> f b -> Bool

liftEq2Const :: Bool -> (a -> b -> Bool) -> (c -> d -> Bool) -> f a c -> f b d -> Bool

compareConst :: Ordering -> a -> a -> Ordering

ltConst :: Bool -> a -> a -> Bool

compare1Const :: Ordering -> f a -> f a -> Ordering

liftCompareConst :: Ordering -> (a -> b -> Ordering) -> f a -> f b -> Ordering

liftCompare2Const :: Ordering -> (a -> b -> Ordering) -> (c -> d -> Ordering) -> f a c -> f b d -> Ordering

readsPrecConst :: ReadS a -> Int -> ReadS a

readPrecConst :: ReadPrec a -> ReadPrec a

readsPrec1Const :: ReadS (f a) -> Int -> ReadS (f a)

liftReadsPrecConst :: ReadS (f a) -> (Int -> ReadS a) -> ReadS [a] -> Int -> ReadS (f a)

liftReadPrecConst :: ReadPrec (f a) -> ReadPrec a -> ReadPrec [a] -> ReadPrec (f a)

liftReadsPrec2Const :: ReadS (f a b) -> (Int -> ReadS a) -> ReadS [a] -> (Int -> ReadS b) -> ReadS [b] -> Int -> ReadS (f a b)

liftReadPrec2Const :: ReadPrec (f a b) -> ReadPrec a -> ReadPrec [a] -> ReadPrec b -> ReadPrec [b] -> ReadPrec (f a b)

showsPrecConst :: ShowS -> Int -> a -> ShowS

showsPrec1Const :: ShowS -> Int -> f a -> ShowS

liftShowsPrecConst :: ShowS -> (Int -> a -> ShowS) -> ([a] -> ShowS) -> Int -> f a -> ShowS

liftShowsPrec2Const :: ShowS -> (Int -> a -> ShowS) -> ([a] -> ShowS) -> (Int -> b -> ShowS) -> ([b] -> ShowS) -> Int -> f a b -> ShowS

data StarKindStatus

Whether a type is not of kind *, is of kind *, or is a kind variable.

Constructors

NotKindStar

KindStar

IsKindVar Name

canRealizeKindStar :: Type -> StarKindStatus

Does a Type have kind * or k (for some kind variable k)?

starKindStatusToName :: StarKindStatus -> Maybe Name

Returns Just the kind variable Name of a StarKindStatus if it exists. Otherwise, returns Nothing.

catKindVarNames :: [StarKindStatus] -> [Name]

Concat together all of the StarKindStatuses that are IsKindVar and extract the kind variables' Names out.

class ClassRep a where

Methods

arity :: a -> Int

allowExQuant :: a -> Bool

fullClassName :: a -> Name

classConstraint :: a -> Int -> Maybe Name

buildTypeInstance

Arguments

:: ClassRep a

=> a
The typeclass for which an instance should be derived

-> Name
The type constructor or data family name

-> Cxt
The datatype context

-> [Type]
The types to instantiate the instance with

-> DatatypeVariant
Are we dealing with a data family instance or not

-> Q (Cxt, Type)

deriveConstraint :: ClassRep a => a -> Type -> (Maybe Pred, [Name])

Attempt to derive a constraint on a Type. If successful, return Just the constraint and any kind variable names constrained to *. Otherwise, return Nothing and the empty list.

See Note [Type inference in derived instances] for the heuristics used to come up with constraints.

checkExistentialContext :: ClassRep a => a -> TyVarMap b -> Cxt -> Name -> Q c -> Q c

noConstructorsError :: Q a

The given datatype has no constructors, and we don't know what to do with it.

derivingKindError :: ClassRep a => a -> Name -> Q b

Either the given data type doesn't have enough type variables, or one of the type variables to be eta-reduced cannot realize kind *.

contravarianceError :: Name -> Q a

The last type variable appeared in a contravariant position when deriving Functor.

noFunctionsError :: Name -> Q a

A constructor has a function argument in a derived Foldable or Traversable instance.

etaReductionError :: Type -> Q a

One of the last type variables cannot be eta-reduced (see the canEtaReduce function for the criteria it would have to meet).

datatypeContextError :: Name -> Type -> Q a

The data type has a DatatypeContext which mentions one of the eta-reduced type variables.

existentialContextError :: Name -> Q a

The data type has an existential constraint which mentions one of the eta-reduced type variables.

outOfPlaceTyVarError :: ClassRep a => a -> Name -> Q b

The data type mentions one of the n eta-reduced type variables in a place other than the last nth positions of a data type in a constructor's field.

enumerationError :: String -> Q a

enumerationOrProductError :: String -> Q a

enumerationErrorStr :: String -> String

typeDataError :: Name -> Q a

type TyVarMap a = Map Name (OneOrTwoNames a)

A mapping of type variable Names to their auxiliary function Names.

type TyVarMap1 = TyVarMap One

type TyVarMap2 = TyVarMap Two

data OneOrTwoNames a where

Constructors

OneName :: Name -> OneOrTwoNames One

TwoNames :: Name -> Name -> OneOrTwoNames Two

data One

data Two

interleave :: [a] -> [a] -> [a]

filterByList :: [Bool] -> [a] -> [a]

filterByList takes a list of Bools and a list of some elements and filters out these elements for which the corresponding value in the list of Bools is False. This function does not check whether the lists have equal length.

filterByLists :: [Bool] -> [a] -> [a] -> [a]

filterByLists takes a list of Bools and two lists as input, and outputs a new list consisting of elements from the last two input lists. For each Bool in the list, if it is True, then it takes an element from the former list. If it is False, it takes an element from the latter list. The elements taken correspond to the index of the Bool in its list. For example:

filterByLists [True, False, True, False] "abcd" "wxyz" = "axcz"

This function does not check whether the lists have equal length.

partitionByList :: [Bool] -> [a] -> ([a], [a])

partitionByList takes a list of Bools and a list of some elements and partitions the list according to the list of Bools. Elements corresponding to True go to the left; elements corresponding to False go to the right. For example, partitionByList [True, False, True] [1,2,3] == ([1,3], [2]) This function does not check whether the lists have equal length.

integerE :: Int -> Q Exp

hasKindStar :: Type -> Bool

Returns True if a Type has kind *.

isStarOrVar :: Kind -> Bool

hasKindVarChain :: Int -> Type -> Maybe [Name]

hasKindVarChain n kind Checks if kind is of the form k_0 -> k_1 -> ... -> k_(n-1), where k0, k1, ..., and k_(n-1) can be * or kind variables.

tyKind :: Type -> Kind

If a Type is a SigT, returns its kind signature. Otherwise, return *.

zipWithAndUnzipM :: Monad m => (a -> b -> m (c, d)) -> [a] -> [b] -> m ([c], [d])

zipWith3AndUnzipM :: Monad m => (a -> b -> c -> m (d, e)) -> [a] -> [b] -> [c] -> m ([d], [e])

thd3 :: (a, b, c) -> c

unsnoc :: [a] -> Maybe ([a], a)

isNullaryCon :: ConstructorInfo -> Bool

conArity :: ConstructorInfo -> Int

Returns the number of fields for the constructor.

isProductType :: NonEmpty ConstructorInfo -> Bool

Returns True if it's a datatype with exactly one, non-existential constructor.

isEnumerationType :: NonEmpty ConstructorInfo -> Bool

Returns True if it's a datatype with one or more nullary, non-GADT constructors.

isVanillaCon :: ConstructorInfo -> Bool

Returns False if we're dealing with existential quantification or GADTs.

newNameList :: String -> Int -> Q [Name]

Generate a list of fresh names with a common prefix, and numbered suffixes.

tvbKind :: TyVarBndr_ flag -> Kind

Extracts the kind from a TyVarBndr.

tvbToType :: TyVarBndr_ flag -> Type

Convert a TyVarBndr to a Type.

applyClass :: Name -> Name -> Pred

Applies a typeclass constraint to a type.

createKindChain :: Int -> Kind

canEtaReduce :: [Type] -> [Type] -> Bool

Checks to see if the last types in a data family instance can be safely eta- reduced (i.e., dropped), given the other types. This checks for three conditions:

All of the dropped types are type variables

All of the dropped types are distinct

None of the remaining types mention any of the dropped types

conTToName :: Type -> Name

Extract the Name from a type constructor. If the argument Type is not a type variable, throw an error.

varTToName_maybe :: Type -> Maybe Name

Extract Just the Name from a type variable. If the argument Type is not a type variable, return Nothing.

varTToName :: Type -> Name

Extract the Name from a type variable. If the argument Type is not a type variable, throw an error.

unSigT :: Type -> Type

Peel off a kind signature from a Type (if it has one).

isTyVar :: Type -> Bool

Is the given type a variable?

isInTypeFamilyApp :: [Name] -> Type -> [Type] -> Q Bool

Detect if a Name in a list of provided Names occurs as an argument to some type family. This makes an effort to exclude oversaturated arguments to type families. For instance, if one declared the following type family:

type family F a :: Type -> Type

Then in the type F a b, we would consider a to be an argument to F, but not b.

allDistinct :: Ord a => [a] -> Bool

Are all of the items in a list (which have an ordering) distinct?

This uses Set (as opposed to nub) for better asymptotic time complexity.

mentionsName :: Type -> [Name] -> Bool

Does the given type mention any of the Names in the list?

predMentionsName :: Pred -> [Name] -> Bool

Does an instance predicate mention any of the Names in the list?

applyTy :: Type -> [Type] -> Type

Construct a type via curried application.

applyTyCon :: Name -> [Type] -> Type

Fully applies a type constructor to its type variables.

unapplyTy :: Type -> (Type, [Type])

Split an applied type into its individual components. For example, this:

Either Int Char

would split to this:

[Either, Int, Char]

uncurryTy :: Type -> (Cxt, [Type])

Split a type signature by the arrows on its spine. For example, this:

forall a b. (a ~ b) => (a -> b) -> Char -> ()

would split to this:

(a ~ b, [a -> b, Char, ()])

uncurryKind :: Kind -> [Kind]

Like uncurryType, except on a kind level.

untagExpr :: [(Name, Name)] -> Q Exp -> Q Exp

tag2ConExpr :: Type -> Q Exp

primOrdFunTbl :: Map Name (Name, Name, Name, Name, Name)

removeClassApp :: Type -> Type

freshen :: Name -> Q Name

freshenType :: Type -> Q Type

enumFromToExpr :: Q Exp -> Q Exp -> Q Exp

primOpAppExpr :: Q Exp -> Name -> Q Exp -> Q Exp

isNonUnitTuple :: Name -> Bool

Checks if a Name represents a tuple type constructor (other than ())

isNonUnitTupleString :: String -> Bool

Checks if a String represents a tuple (other than ())

isInfixDataCon :: String -> Bool

Checks if a String names a valid Haskell infix data constructor (i.e., does it begin with a colon?).

isSym :: String -> Bool

ghc7'8OrLater :: Bool

isTrueHashValName :: Name

fmapConstValName :: Name

replaceConstValName :: Name

foldrConstValName :: Name

foldMapConstValName :: Name

nullConstValName :: Name

traverseConstValName :: Name

eqConstValName :: Name

eq1ConstValName :: Name

liftEqConstValName :: Name

liftEq2ConstValName :: Name

compareConstValName :: Name

ltConstValName :: Name

compare1ConstValName :: Name

liftCompareConstValName :: Name

liftCompare2ConstValName :: Name

readsPrecConstValName :: Name

readPrecConstValName :: Name

readsPrec1ConstValName :: Name

liftReadsPrecConstValName :: Name

liftReadPrecConstValName :: Name

liftReadsPrec2ConstValName :: Name

liftReadPrec2ConstValName :: Name

showsPrecConstValName :: Name

showsPrec1ConstValName :: Name

liftShowsPrecConstValName :: Name

liftShowsPrec2ConstValName :: Name

viaTypeName :: Name

cHashDataName :: Name

dHashDataName :: Name

fHashDataName :: Name

identDataName :: Name

iHashDataName :: Name

puncDataName :: Name

symbolDataName :: Name

wrapMonadDataName :: Name

addrHashTypeName :: Name

boundedTypeName :: Name

charHashTypeName :: Name

doubleHashTypeName :: Name

enumTypeName :: Name

floatHashTypeName :: Name

foldableTypeName :: Name

functorTypeName :: Name

intTypeName :: Name

intHashTypeName :: Name

ixTypeName :: Name

readTypeName :: Name

showTypeName :: Name

traversableTypeName :: Name

wordHashTypeName :: Name

altValName :: Name

appendValName :: Name

chooseValName :: Name

composeValName :: Name

constValName :: Name

enumFromValName :: Name

enumFromThenValName :: Name

enumFromThenToValName :: Name

enumFromToValName :: Name

eqAddrHashValName :: Name

eqCharHashValName :: Name

eqDoubleHashValName :: Name

eqFloatHashValName :: Name

eqIntHashValName :: Name

eqWordHashValName :: Name

errorValName :: Name

flipValName :: Name

fmapValName :: Name

foldrValName :: Name

foldMapValName :: Name

fromEnumValName :: Name

geAddrHashValName :: Name

geCharHashValName :: Name

geDoubleHashValName :: Name

geFloatHashValName :: Name

geIntHashValName :: Name

getTagValName :: Name

geWordHashValName :: Name

gtAddrHashValName :: Name

gtCharHashValName :: Name

gtDoubleHashValName :: Name

gtFloatHashValName :: Name

gtIntHashValName :: Name

gtWordHashValName :: Name

idValName :: Name

indexValName :: Name

inRangeValName :: Name

leAddrHashValName :: Name

leCharHashValName :: Name

leDoubleHashValName :: Name

leFloatHashValName :: Name

leIntHashValName :: Name

leWordHashValName :: Name

listValName :: Name

ltAddrHashValName :: Name

ltCharHashValName :: Name

ltDoubleHashValName :: Name

ltFloatHashValName :: Name

ltIntHashValName :: Name

ltWordHashValName :: Name

minBoundValName :: Name

mapValName :: Name

maxBoundValName :: Name

minusIntHashValName :: Name

neqIntHashValName :: Name

parenValName :: Name

parensValName :: Name

pfailValName :: Name

plusValName :: Name

precValName :: Name

predValName :: Name

rangeSizeValName :: Name

rangeValName :: Name

readFieldHash :: String -> ReadPrec a -> ReadPrec a

readFieldHashValName :: Name

readListValName :: Name

readListPrecDefaultValName :: Name

readListPrecValName :: Name

readPrec_to_SValName :: Name

readPrecValName :: Name

readS_to_PrecValName :: Name

readsPrecValName :: Name

replaceValName :: Name

resetValName :: Name

returnValName :: Name

seqValName :: Name

showCharValName :: Name

showListValName :: Name

showListWithValName :: Name

showParenValName :: Name

showsPrecValName :: Name

showSpaceValName :: Name

showStringValName :: Name

stepValName :: Name

succValName :: Name

tagToEnumHashValName :: Name

timesValName :: Name

toEnumValName :: Name

traverseValName :: Name

unsafeIndexValName :: Name

unsafeRangeSizeValName :: Name

unwrapMonadValName :: Name

boolTypeName :: Name

falseDataName :: Name

trueDataName :: Name

eqDataName :: Name

gtDataName :: Name

ltDataName :: Name

eqTypeName :: Name

ordTypeName :: Name

andValName :: Name

compareValName :: Name

eqValName :: Name

geValName :: Name

gtValName :: Name

leValName :: Name

ltValName :: Name

notValName :: Name

wHashDataName :: Name

expectPValName :: Name

allValName :: Name

apValName :: Name

pureValName :: Name

liftA2ValName :: Name

mappendValName :: Name

memptyValName :: Name

nullValName :: Name

eq1TypeName :: Name

ord1TypeName :: Name

read1TypeName :: Name

show1TypeName :: Name

eq2TypeName :: Name

ord2TypeName :: Name

read2TypeName :: Name

show2TypeName :: Name

liftEqValName :: Name

liftEq2ValName :: Name

liftCompareValName :: Name

liftCompare2ValName :: Name

liftReadsPrecValName :: Name

liftReadListValName :: Name

liftReadsPrec2ValName :: Name

liftReadList2ValName :: Name

liftShowListValName :: Name

liftShowsPrecValName :: Name

liftShowList2ValName :: Name

liftShowsPrec2ValName :: Name

coerceValName :: Name

liftReadListPrecDefaultValName :: Name

liftReadListPrec2DefaultValName :: Name

liftReadListPrecValName :: Name

liftReadListPrec2ValName :: Name

liftReadPrecValName :: Name

liftReadPrec2ValName :: Name

showCommaSpaceValName :: Name

appEndoValName :: Name

dualDataName :: Name

endoDataName :: Name

getDualValName :: Name

readFieldValName :: Name

readSymFieldValName :: Name

eqInt8HashValName :: Name

eqInt16HashValName :: Name

eqWord8HashValName :: Name

eqWord16HashValName :: Name

geInt8HashValName :: Name

geInt16HashValName :: Name

geWord8HashValName :: Name

geWord16HashValName :: Name

gtInt8HashValName :: Name

gtInt16HashValName :: Name

gtWord8HashValName :: Name

gtWord16HashValName :: Name

int8HashTypeName :: Name

int8ToIntHashValName :: Name

int16HashTypeName :: Name

int16ToIntHashValName :: Name

intToInt8HashValName :: Name

intToInt16HashValName :: Name

leInt8HashValName :: Name

leInt16HashValName :: Name

leWord8HashValName :: Name

leWord16HashValName :: Name

ltInt8HashValName :: Name

ltInt16HashValName :: Name

ltWord8HashValName :: Name

ltWord16HashValName :: Name

word8HashTypeName :: Name

word8ToWordHashValName :: Name

word16HashTypeName :: Name

word16ToWordHashValName :: Name

wordToWord8HashValName :: Name

wordToWord16HashValName :: Name

eqInt32HashValName :: Name

eqWord32HashValName :: Name

geInt32HashValName :: Name

geWord32HashValName :: Name

gtInt32HashValName :: Name

gtWord32HashValName :: Name

int32HashTypeName :: Name

int32ToIntHashValName :: Name

intToInt32HashValName :: Name

leInt32HashValName :: Name

leWord32HashValName :: Name

ltInt32HashValName :: Name

ltWord32HashValName :: Name

word32HashTypeName :: Name

word32ToWordHashValName :: Name

wordToWord32HashValName :: Name