Data.Enum.Deriving.Internal

Description

Exports functions to mechanically derive Enum instances.

Note: this is an internal module, and as such, the API presented here is not guaranteed to be stable, even between minor releases of this library.

Enum

deriveEnum :: Name -> Q [Dec]

Generates an Enum instance declaration for the given data type or data family instance.

makeSucc :: Name -> Q Exp

Generates a lambda expression which behaves like succ (without requiring an Enum instance).

makePred :: Name -> Q Exp

Generates a lambda expression which behaves like pred (without requiring an Enum instance).

makeToEnum :: Name -> Q Exp

Generates a lambda expression which behaves like toEnum (without requiring an Enum instance).

makeFromEnum :: Name -> Q Exp

Generates a lambda expression which behaves like fromEnum (without requiring an Enum instance).

makeEnumFrom :: Name -> Q Exp

Generates a lambda expression which behaves like enumFrom (without requiring an Enum instance).

makeEnumFromThen :: Name -> Q Exp

Generates a lambda expression which behaves like enumFromThen (without requiring an Enum instance).