Data.Functor.Deriving

Description

Exports functions to mechanically derive Functor instances.

For more info on how deriving Functor works, see this GHC wiki page.

Functor

deriveFunctor :: Name -> Q [Dec]

Generates a FunctorClass instance declaration for the given data type or data family instance.

deriveFunctorOptions :: FFTOptions -> Name -> Q [Dec]

Like deriveFunctor, but takes an FFTOptions argument.

makeFmap :: Name -> Q Exp

Generates a lambda expression which behaves like fmap (without requiring a FunctorClass instance).

makeFmapOptions :: FFTOptions -> Name -> Q Exp

Like makeFmap, but takes an FFTOptions argument.

makeReplace :: Name -> Q Exp

Generates a lambda expression which behaves like (<$) (without requiring a FunctorClass instance).

makeReplaceOptions :: FFTOptions -> Name -> Q Exp

Like makeReplace, but takes an FFTOptions argument.

FFTOptions

newtype FFTOptions

Options that further configure how the functions in Data.Functor.Deriving should behave. (FFT stands for FunctorClassFunctorClassFunctorClass.)

Constructors

FFTOptions

Fields

fftEmptyCaseBehavior :: Bool

If True, derived instances for empty data types (i.e., ones with no data constructors) will use the EmptyCase language extension. If False, derived instances will simply use seq instead. (This has no effect on GHCs before 7.8, since EmptyCase is only available in 7.8 or later.)

defaultFFTOptions :: FFTOptions

Conservative FFTOptions that doesn't attempt to use EmptyCase (to prevent users from having to enable that extension at use sites.)

deriveFunctor limitations

Be aware of the following potential gotchas:

Type variables of kind * -> * are assumed to have Functor constraints. If this is not desirable, use makeFmap.