Data.Ord.Deriving.Internal

Description

Exports functions to mechanically derive OrdClass, Ord1, and Ord2 instances.

Note: this is an internal module, and as such, the API presented here is not guaranteed to be stable, even between minor releases of this library.

OrdClass

deriveOrd :: Name -> Q [Dec]

Generates an OrdClass instance declaration for the given data type or data family instance.

makeCompare :: Name -> Q Exp

Generates a lambda expression which behaves like compare (without requiring an OrdClass instance).

makeLE :: Name -> Q Exp

Generates a lambda expression which behaves like (<=) (without requiring an OrdClass instance).

makeLT :: Name -> Q Exp

Generates a lambda expression which behaves like (<) (without requiring an OrdClass instance).

makeGT :: Name -> Q Exp

Generates a lambda expression which behaves like (>) (without requiring an OrdClass instance).

makeGE :: Name -> Q Exp

Generates a lambda expression which behaves like (>=) (without requiring an OrdClass instance).

makeMax :: Name -> Q Exp

Generates a lambda expression which behaves like max (without requiring an OrdClass instance).

makeMin :: Name -> Q Exp

Generates a lambda expression which behaves like min (without requiring an OrdClass instance).

Ord1

deriveOrd1 :: Name -> Q [Dec]

Generates an Ord1 instance declaration for the given data type or data family instance.

makeLiftCompare :: Name -> Q Exp

Generates a lambda expression which behaves like liftCompare (without requiring an Ord1 instance).

This function is not available with transformers-0.4.

makeCompare1 :: Name -> Q Exp

Generates a lambda expression which behaves like compare1 (without requiring an Ord1 instance).

Ord2

deriveOrd2 :: Name -> Q [Dec]

Generates an Ord2 instance declaration for the given data type or data family instance.

This function is not available with transformers-0.4.

makeLiftCompare2 :: Name -> Q Exp

Generates a lambda expression which behaves like liftCompare2 (without requiring an Ord2 instance).

This function is not available with transformers-0.4.

makeCompare2 :: Name -> Q Exp

Generates a lambda expression which behaves like compare2 (without requiring an Ord2 instance).

This function is not available with transformers-0.4.