Data.Ix.Deriving

Description

Exports functions to mechanically derive Ix instances.

Ix

deriveIx :: Name -> Q [Dec]

Generates a Ix instance declaration for the given data type or data family instance.

makeRange :: Name -> Q Exp

Generates a lambda expression which behaves like range (without requiring an Ix instance).

makeUnsafeIndex :: Name -> Q Exp

Generates a lambda expression which behaves like unsafeIndex (without requiring an Ix instance).

makeInRange :: Name -> Q Exp

Generates a lambda expression which behaves like inRange (without requiring an Ix instance).

deriveIx limitations

Be aware of the following potential gotchas:

Type variables of kind * are assumed to have Ix constraints. If this is not desirable, use makeRange or one of its cousins.